Keyboard accessibility for sortable header cells
Users can alter the tbody cell rendering to add accessibility features, but the header row is not customizable. This allows end users to sort using keyboard controls.

If English text on the IconArrowUp ariaLabel is undesirable in the interest of keeping the component language-independent, ↑ and ↓ can be used for a slightly degraded experience; this is fine for anything converting it to text, but I've confirmed that screen readers still say "up arrow" and "down arrow" for these glyphs regardless of the language selected, so I'm not sure it's ultimately any better.