🚑️ removes the default storeColumnKey
The default value created an infinite loop when having multiple DataTables on the same view. This was because using the same key, the data was always read and written to the same storage. Now, if the key is undefined, no data is written, indicating that the resizing, sizing, or sorting functions on the columns are not being used.