feat: Rudementary inline cell editing as a library function
Implement simple cell editing that edits the record from the datatable
data in place. Provide the onEdit callback to allow hooking external
logic to the featureset.

This is mostly a small version of the grander feature as has been
discussed at length in some issues where a workaround was suggested to
use the render function of the column. However, this appears to be very
cumbersome to do and maintain. In this case, I belive that I speak for
others too, that this may be a good core feature of the datatable with
room for expansion in the future.